html/webappapis/dynamic-markup-insertion/document-write/module-tla-import.html

<!doctype html>
<title>document.write in an imported module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const iframe = document.createElement("iframe");

  iframe.onerror = t.unreached_func("Error loading iframe");
  let testEndWasCalled = false;
  document.addEventListener("testEnd", t.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, "PASS\n");
    testEndWasCalled = true;
  }));
  iframe.onload = t.step_func_done(() => {
    assert_true(testEndWasCalled, 'testEnd event was not sent');
    assert_equals(iframe.contentDocument.body.textContent, "PASS\n");
  });

  iframe.src = "module-tla-import-iframe.html";
  document.body.appendChild(iframe);
});
</script>
